Refactored data domain
