+++ cwe_checker_rs/src/abstract_domain/mod.rs
@@ -12,6 +12,9 @@ pub use identifier::*;
 mod pointer;
 pub use pointer::*;
 
+mod data;
+pub use data::*;
+
 /// The main trait describing an abstract domain.
 ///
 /// Each abstract domain is partially ordered and has a maximal element (which can be generated by `top()`).
