+++ cwe_checker_rs/src/analysis/pointer_inference/context.rs
@@ -7,8 +7,8 @@ use crate::term::*;
 use crate::utils::log::*;
 use std::collections::{BTreeMap, BTreeSet, HashSet};
 
-use super::data::Data;
 use super::state::State;
+use super::Data;
 
 pub struct Context<'a> {
     pub graph: Graph<'a>,
@@ -310,7 +310,7 @@ impl<'a> crate::analysis::interprocedural_fixpoint::Problem<'a> for Context<'a> 
             self.log_debug(
                 new_state.set_register(
                     stack_register,
-                    stack_pointer.bin_op(crate::bil::BinOpType::PLUS, &Data::bitvector(offset)),
+                    stack_pointer.bin_op(crate::bil::BinOpType::PLUS, &offset.into()),
                 ),
                 Some(&call.tid),
             );
@@ -596,7 +596,7 @@ mod tests {
     #[test]
     fn context_problem_implementation() {
         use crate::analysis::interprocedural_fixpoint::Problem;
-        use crate::analysis::pointer_inference::data::*;
+        use crate::analysis::pointer_inference::Data;
         use crate::bil::*;
         use Expression::*;
 
@@ -764,8 +764,8 @@ mod tests {
     #[test]
     fn update_return() {
         use crate::analysis::interprocedural_fixpoint::Problem;
-        use crate::analysis::pointer_inference::data::*;
         use crate::analysis::pointer_inference::object::ObjectType;
+        use crate::analysis::pointer_inference::Data;
         let project = mock_project();
         let (cwe_sender, _cwe_receiver) = crossbeam_channel::unbounded();
         let (log_sender, _log_receiver) = crossbeam_channel::unbounded();
